+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### [Unreleased]
+### Added
+- Inventory containers.
 ### Fixed
 - Weapon profiles were not generated for classless Character Actors.